Attempting to solve UnboundExecutionError.
It seems that this problem occurs because the unit test is executing outside
the Cherrypy server's context. Hence, DBOperations in unit test is failing.

Trying to run a common object with the unit test and the Cherrypy server.
Still not succeeding (due to the class structures of unittest.TestCase and
Cherrypy's server).